RepairFuelGene fixed
There are some problem when the gene values goes upper than upper limit.
The problem is fixed now using the concept of rescaling those value
according to the maximum value of fuels share. If the maximum value of
any share have  a value larger than 1.00, we will scaling back to 1.00.